support both major versions of fernet
